+++ packages/sanity/src/core/form/inputs/PortableText/object/modals/ObjectEditModal.tsx
@@ -72,7 +72,7 @@ export function ObjectEditModal(props: {
       type="dialog"
       onClose={onClose}
       header={modalTitle}
-      width={1}
+      width={modalWidth ?? 1}
       autofocus={autoFocus}
     >
       {props.children}