+++ packages/@sanity/types/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+## [5.31.0](https://github.com/sanity-io/sanity/compare/v5.30.0...v5.31.0) (2026-06-10)
+
+### Bug Fixes
+
+* **core:** support modal width in Enhanced dialog ([#12955](https://github.com/sanity-io/sanity/issues/12955)) ([37d9ece](https://github.com/sanity-io/sanity/commit/37d9ece0b9ca504b1fa51f5d49df0c7d36ba865c))
+* **deps:** update dependency @sanity/insert-menu to v3.0.8 ([#12941](https://github.com/sanity-io/sanity/issues/12941)) ([dd02d30](https://github.com/sanity-io/sanity/commit/dd02d30b27c5154c3665e35e2019633a6e2c4104))
 ## [5.28.0](https://github.com/sanity-io/sanity/compare/v5.27.0...v5.28.0) (2026-05-28)
 
 ### Bug Fixes